Add PBX sync scripts, migrate to HCP project
Add a suite of scripts for syncing and reconciling Xcode PBX/SSK sources (sync_pbx_from_signal.sh, sync_pbx_file_refs_from_signal.py, reconcile_ssk_swift_sources.py, prune_ssk_missing_objc.py, fix_hcp_pbx_swift_paths.py, etc.) and update the HCP.xcodeproj project file. Remove Signal.xcodeproj workspace/shared scheme/baseline files and introduce supporting files including SignalUI-Bridging-Header.h, WebAppTabPinsStore.swift, SSOStrings.swift, and CursorAgentDebugNDJSON.swift. Numerous Signal, SignalServiceKit and SignalUI source files were updated to align with the project reorganization and feature updates. Also update Podfile and build-and-test script to reflect the new project layout and build workflow.

+++ SignalUI/FormatStyles/OWSPercentFormatStyle.swift
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 //
 
+import Foundation
 public struct OWSPercentFormatStyle: FormatStyle {
     private let fractionDigits: Int
 
@@ -10,7 +11,7 @@ public struct OWSPercentFormatStyle: FormatStyle {
         self.fractionDigits = fractionDigits
     }
 
-    public func format(_ percent: Float) -> String {
+    public func format(_ percent: Double) -> String {
         return percent.formatted(.percent.precision(.fractionLength(fractionDigits)))
     }
 }
